+++ ts-client/index.ts
@@ -1,6 +1,6 @@
-export * as constants from './src/constants';
+export * as constants from "./src/constants/vault";
 export * as utils from "./src/utils";
-export * as vault from "./src/vault";
+export * as vault from "./src/model/vault/vault";
 
-import Vault from "./src/vault";
+import Vault from "./src/model/vault/vault";
 export default Vault;
