+++ ts-client/src/model/vault/strategy/mango.ts
@@ -2,13 +2,13 @@ import * as mango from "@blockworks-foundation/mango-client";
 import * as anchor from "@project-serum/anchor";
 import { TOKEN_PROGRAM_ID } from "@solana/spl-token";
 import {
-  AccountMeta, PublicKey, TransactionInstruction
+  AccountMeta,
+  PublicKey,
+  TransactionInstruction,
 } from "@solana/web3.js";
 
-import { StrategyHandler } from ".";
-import { SEEDS } from "../constants";
-import { Strategy } from "../mint";
-import { VaultProgram } from "../vault";
+import { SEEDS } from "../../../constants/vault";
+import { Strategy, StrategyHandler, VaultProgram } from "../../../types/vault";
 
 export default class MangoHandler implements StrategyHandler {
   static MangoProgramId = new PublicKey(
@@ -55,13 +55,13 @@ export default class MangoHandler implements StrategyHandler {
     );
 
     const rootBankState = mangoGroupState.rootBankAccounts[rootBankIdx];
-    if (!rootBankState) return '';
+    if (!rootBankState) return "";
     const nodeBankPK = rootBankState.nodeBanks[0];
 
     const nodeBankState = rootBankState.nodeBankAccounts.find(
       (t) => t.publicKey.toBase58() === nodeBankPK.toBase58()
     );
-    if (!nodeBankState) return '';
+    if (!nodeBankState) return "";
     const accountData = [
       { pubkey: MangoHandler.MangoGrouPK, isWritable: true },
       { pubkey: mangoAccountPK, isWritable: true },
