+++ ts-client/src/model/vault/strategy/solendWithoutLM.ts
@@ -1,34 +1,40 @@
 import {
-  PublicKey,
   AccountMeta,
+  PublicKey,
   SYSVAR_CLOCK_PUBKEY,
-  Cluster,
   TransactionInstruction,
 } from "@solana/web3.js";
+import * as solend from "@solendprotocol/solend-sdk";
 import { TOKEN_PROGRAM_ID } from "@solana/spl-token";
-
 import * as anchor from "@project-serum/anchor";
 
-import * as port from "@port.finance/port-sdk";
+import { SEEDS } from "../../../constants/vault";
+import {
+  ReserveState,
+  Strategy,
+  StrategyHandler,
+  VaultProgram,
+} from "../../../types/vault";
 
-import { ReserveState, StrategyHandler } from ".";
-import { VaultProgram } from "../vault";
-import { SEEDS } from "../constants";
-import { Strategy } from "../mint";
-
-export default class PortWithoutLMHandler implements StrategyHandler {
+export default class SolendWithoutLMHandler implements StrategyHandler {
   constructor(public strategyProgram: PublicKey) {}
+
   async getReserveState(
     program: VaultProgram,
     reserve: PublicKey
   ): Promise<ReserveState> {
-    const account = await program.provider.connection.getAccountInfo(reserve);
-    const state = port.ReserveLayout.decode(account!.data) as port.ReserveData;
+    const state = await (async () => {
+      const account = await program.provider.connection.getAccountInfo(reserve);
+
+      const solendParse = solend.parseReserve(account!.owner, account!);
+      return solendParse!.info;
+    })();
+
     return {
       collateral: {
         mintPubkey: state.collateral.mintPubkey,
-        mintTotalSupply: Number(state.collateral.mintTotalSupply.toU64()),
-        supplyPubkey: state.collateral.supplyPubkey.toString(),
+        mintTotalSupply: state.collateral.mintTotalSupply.toNumber(),
+        supplyPubkey: state.collateral.mintTotalSupply.toString(),
       },
       state,
     };
@@ -48,44 +54,48 @@ export default class PortWithoutLMHandler implements StrategyHandler {
     preInstructions: TransactionInstruction[],
     postInstructions: TransactionInstruction[]
   ): Promise<string> {
-    const { state } = await this.getReserveState(
+    const { collateral, state } = await this.getReserveState(
       program,
       strategy.state.reserve
     );
-    const strategyBuffer = new PublicKey(strategy.pubkey).toBuffer();
 
     let [collateralVault] = await PublicKey.findProgramAddress(
-      [Buffer.from(SEEDS.COLLATERAL_VAULT_PREFIX), strategyBuffer],
+      [
+        Buffer.from(SEEDS.COLLATERAL_VAULT_PREFIX),
+        new PublicKey(strategy.pubkey).toBuffer(),
+      ],
       program.programId
     );
 
-    const {
-      collateral: portCollateral,
-      lendingMarket,
-      liquidity,
-    } = state as port.ReserveData;
+    const { liquidity, lendingMarket } = state as solend.Reserve;
 
     const [lendingMarketAuthority] = await PublicKey.findProgramAddress(
       [lendingMarket.toBuffer()],
       this.strategyProgram
     );
 
-    const accountData = [
-      { pubkey: portCollateral.supplyPubkey, isWritable: true },
+    const accounts = [
+      { pubkey: liquidity.supplyPubkey, isWritable: true },
       { pubkey: lendingMarket },
       { pubkey: lendingMarketAuthority },
-      { pubkey: portCollateral.mintPubkey, isWritable: true },
+      { pubkey: collateral.mintPubkey, isWritable: true },
       { pubkey: SYSVAR_CLOCK_PUBKEY },
     ];
     const remainingAccounts: Array<AccountMeta> = [];
-    for (const account of accountData) {
+    for (const account of accounts) {
       remainingAccounts.push({
         pubkey: account.pubkey,
         isWritable: !!account.isWritable,
         isSigner: false,
       });
     }
-
+    const { pythOracle, switchboardOracle } = liquidity as {
+      pythOracle?: PublicKey;
+      switchboardOracle?: PublicKey;
+    };
+    if (!pythOracle || !switchboardOracle) {
+      throw Error("Cannot get pythOracle or switchboardOracle from Solend");
+    }
     const tx = await program.methods
       .withdrawDirectlyFromStrategy(new anchor.BN(amount), new anchor.BN(0))
       .accounts({
@@ -105,16 +115,16 @@ export default class PortWithoutLMHandler implements StrategyHandler {
       .remainingAccounts(remainingAccounts)
       .preInstructions(
         preInstructions.concat([
-          port.refreshReserveInstruction(
+          solend.refreshReserveInstruction(
             strategy.state.reserve,
-            liquidity.oraclePubkey,
-            this.strategyProgram
+            this.strategyProgram,
+            pythOracle,
+            switchboardOracle
           ),
         ])
       )
       .postInstructions(postInstructions)
       .rpc();
-
     return tx;
   }
 }
