+++ ts-client/src/model/vault/strategy/vault.ts
@@ -1,11 +1,7 @@
-import { PublicKey, TransactionInstruction } from '@solana/web3.js';
-import { TOKEN_PROGRAM_ID } from '@solana/spl-token';
-
-import * as anchor from '@project-serum/anchor';
-
-import { VaultProgram } from '../vault';
-import { StrategyHandler } from '.';
-import { Strategy } from '../mint';
+import { PublicKey, TransactionInstruction } from "@solana/web3.js";
+import { TOKEN_PROGRAM_ID } from "@solana/spl-token";
+import * as anchor from "@project-serum/anchor";
+import { Strategy, StrategyHandler, VaultProgram } from "../../../types/vault";
 
 export default class VaultHandler implements StrategyHandler {
   async withdraw(
@@ -20,7 +16,7 @@ export default class VaultHandler implements StrategyHandler {
     userLp: PublicKey,
     amount: number,
     preInstructions: TransactionInstruction[],
-    postInstructions: TransactionInstruction[],
+    postInstructions: TransactionInstruction[]
   ): Promise<string> {
     const tx = await program.methods
       .withdraw(new anchor.BN(amount), new anchor.BN(0))
