+++ ts-client/src/types/clock_state.ts
