--- ts-client/types/program_result.ts
@@ -1,11 +1,0 @@
-export interface ProgramResult {
-  context: {
-    slot: number;
-  };
-  value: {
-    amount: string;
-    decimals: number;
-    uiAmount: number;
-    uiAmountString: string;
-  };
-}
